a2a-multi-agent-keda-scaling/pyproject.toml
[project]
name = "a2a-multi-agent-keda-scaling"
version = "0.1.0"
description = "Multi-agent demo with A2A Protocol, LangGraph, KEDA scaling on AKS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.1",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
    "starlette>=0.46.1",
    "sse-starlette>=2.2.1",
    "azure-servicebus>=7.12.0",
    "azure-identity>=1.15.0",
    "prometheus-client>=0.20.0",
    "aiohttp>=3.13.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["agents", "common"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"
